airflow/requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.26.4
scikit-learn==1.3.2
psycopg2-binary==2.9.9
mlflow==2.16.0
xgboost==2.0.3
matplotlib==3.7.5
seaborn==0.12.2
pyarrow==16.1.0
torch==2.3.1
pyspark==3.5.0
apache-airflow-providers-apache-spark>=4.0.0,<5.0.0
